reorganize: library internals are hidden behind interface
