+++ src/stores/library.ts
@@ -2,8 +2,7 @@
  * Svelte store to wrap the Library client.
  * See {@link initClient} for more details about setting up the client.
  */
-import { initClient } from "@/lib/services/library/libraryCommsManager";
-import { type Library, type Options } from "@/lib/services/library/_types";
+import { initClient, type Library, type Options } from "@/lib/services/library";
 import { get, writable } from "svelte/store";
 
 let resolveClientReady: () => void;
@@ -38,4 +37,4 @@ export const initLibrary = async (options: Options) => {
 
 export const shutdownLibrary = () => {
 	libraryClientStore.set(null as unknown as Library);
-}
+};
